examples/dqmanalysis/requirements.txt
kafka-python==2.0.2
numpy==1.19.5
psycopg2-binary==2.9.1
pymongo==3.12.0
